refactor(api): move test in prescription/campaign/acceptance/application/campaign-administration-route_test.js
